chore: Move view base classes to its own file
